Improve memory safety
This commit makes numerous refinements to cosmopolitan memory handling.

The default stack size has been reduced from 2mb to 128kb. A new macro
is now provided so you can easily reconfigure the stack size to be any
value you want. Work around the breaking change by adding to your main:

    STATIC_STACK_SIZE(0x00200000);  // 2mb stack

If you're not sure how much stack you need, then you can use:

    STATIC_YOINK("stack_usage_logging");

After which you can `sort -nr o/$MODE/stack.log`. Based on the unit test
suite, nothing in the Cosmopolitan repository (except for Python) needs
a stack size greater than 30kb. There are also new macros for detecting
the size and address of the stack at runtime, e.g. GetStackAddr(). We
also now support sigaltstack() so if you want to see nice looking crash
reports whenever a stack overflow happens, you can put this in main():

    ShowCrashReports();

Under `make MODE=dbg` and `make MODE=asan` the unit testing framework
will now automatically print backtraces of memory allocations when
things like memory leaks happen. Bugs are now fixed in ASAN global
variable overrun detection. The memtrack and asan runtimes also handle
edge cases now. The new tools helped to identify a few memory leaks,
which are fixed by this change.

This change should fix an issue reported in #288 with ARG_MAX limits.
Fixing this doubled the performance of MKDEPS.COM and AR.COM yet again.

+++ tool/build/mkdeps.c
@@ -45,6 +45,7 @@
 #include "libc/sysv/consts/prot.h"
 #include "libc/x/x.h"
 #include "third_party/getopt/getopt.h"
+#include "tool/build/lib/getargs.h"
 
 #define MAX_READ FRAMESIZE
 
@@ -223,59 +224,51 @@ wontreturn void OnMissingFile(const char *list, const char *src) {
    * notation on mkdeps related rules, the build will
    * automatically restart itself.
    */
-  fprintf(stderr, "%s %s...\n", "Refreshing", list);
-  unlink(list);
+  if (list) {
+    fprintf(stderr, "%s %s...\n", "Refreshing", list);
+    unlink(list);
+  }
   exit(1);
 }
 
 void LoadRelationships(int argc, char *argv[]) {
   int fd;
+  char *buf;
   ssize_t rc;
   bool skipme;
-  FILE *finpaths;
   struct Edge edge;
-  char *line, *buf;
+  struct GetArgs ga;
   unsigned srcid, dependency;
-  size_t i, linecap, inclen, size;
+  size_t i, inclen, size;
   const char *p, *pe, *src, *path, *pathend;
-  line = NULL;
-  linecap = 0;
   inclen = strlen(kIncludePrefix);
   buf = gc(xmemalign(PAGESIZE, PAGESIZE + MAX_READ + 16));
   buf += PAGESIZE;
   buf[-1] = '\n';
-  for (i = optind; i < argc; ++i) {
-    if (!(finpaths = fopen(argv[i], "r"))) {
-      fprintf(stderr, "\n\e[1mERROR: %s FAILED BECAUSE %s CAUSED %m\e[0m\n\n",
-              argv[0], argv[i]);
-      exit(1);
-    }
-    while (getline(&line, &linecap, finpaths) != -1) {
-      src = chomp(line);
-      if (ShouldSkipSource(src)) continue;
-      srcid = GetSourceId(src, strlen(src));
-      if ((fd = open(src, O_RDONLY)) == -1) OnMissingFile(argv[i], src);
-      CHECK_NE(-1, (rc = read(fd, buf, MAX_READ)));
-      close(fd);
-      size = rc;
-      bzero(buf + size, 16);
-      for (p = buf, pe = p + size; p < pe; ++p) {
-        p = strstr(p, kIncludePrefix);
-        if (!p) break;
-        path = p + inclen;
-        pathend = memchr(path, '"', pe - path);
-        if (pathend && (p[-1] == '#' || p[-1] == '.') && p[-2] == '\n') {
-          dependency = GetSourceId(path, pathend - path);
-          edge.from = srcid;
-          edge.to = dependency;
-          append(&edges, &edge);
-          p = pathend;
-        }
+  getargs_init(&ga, argv + optind);
+  while ((src = getargs_next(&ga))) {
+    if (ShouldSkipSource(src)) continue;
+    srcid = GetSourceId(src, strlen(src));
+    if ((fd = open(src, O_RDONLY)) == -1) OnMissingFile(ga.path, src);
+    CHECK_NE(-1, (rc = read(fd, buf, MAX_READ)));
+    close(fd);
+    size = rc;
+    bzero(buf + size, 16);
+    for (p = buf, pe = p + size; p < pe; ++p) {
+      p = strstr(p, kIncludePrefix);
+      if (!p) break;
+      path = p + inclen;
+      pathend = memchr(path, '"', pe - path);
+      if (pathend && (p[-1] == '#' || p[-1] == '.') && p[-2] == '\n') {
+        dependency = GetSourceId(path, pathend - path);
+        edge.from = srcid;
+        edge.to = dependency;
+        append(&edges, &edge);
+        p = pathend;
       }
     }
-    CHECK_NE(-1, fclose(finpaths));
   }
-  free(line);
+  getargs_destroy(&ga);
 }
 
 void GetOpts(int argc, char *argv[]) {
